+++ ta-gui/src/app/aluno.service.ts
@@ -21,7 +21,7 @@ export class AlunoService {
   }
 
   atualizar(aluno: Aluno): Observable<Aluno> {
-    return this.http.put<any>(this.taURL + "/aluno",JSON.stringify(aluno), {headers: this.headers})
+    return this.http.put<any>(this.taURL + "/aluno", JSON.stringify(aluno), {headers: this.headers})
       .pipe( 
         retry(2),
         map( res => {if (res.success) {return aluno;} else {return null;}} )
