+++ ta-gui/src/app/cadeiras.service.ts
@@ -1,173 +1,116 @@
 import { Injectable } from '@angular/core';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Observable } from 'rxjs';
+import { retry, map } from 'rxjs/operators';
+
 import { Aluno } from '..//../../common/aluno';
 import { Cadeira } from "../../../common/cadeiras";
 
+export class CadeiraAluno {
+    cadeira = new Cadeira();
+    aluno = new Aluno();
+
+    constructor(cadeira: Cadeira, aluno: Aluno) {
+        this.cadeira = cadeira;
+        this.aluno = aluno;
+    }
+}
+
 @Injectable()
 export class CadeiraService {
-  cadeiras: Cadeira[] = [];
-  departamentos: Set<string> = new Set<string>(["CAC",
-                                                "CB",
-                                                "CCEN",
-                                                "CCJ",
-                                                "CCS",
-                                                "CCM",
-                                                "CCSA",
-                                                "CE",
-                                                "CFCH",
-                                                "CIn",
-                                                "CTG"]);
-
-  criar(cadeira: Cadeira): Cadeira | string {
-    cadeira = cadeira.clone();
-    var result_check = this.checkCriar(cadeira);
-    if (result_check == "ok") {
-      this.cadeiras.push(cadeira);
-      if (!this.departamentos.has(cadeira.departamento_ofertante)) {
-        this.departamentos.add(cadeira.departamento_ofertante);
-      }
-      return cadeira;
+    private headers = new HttpHeaders({'Content-Type': 'application/json'});
+    private taURL = 'http://localhost:3000';
+    
+    constructor(private http: HttpClient) {}
+    criar(cadeira: Cadeira): Observable<Cadeira> {
+        return this.http.post<any>(this.taURL + "/cadeira", cadeira, {headers: this.headers})
+            .pipe(
+                retry(2),
+                map( res => {
+                    if (res.sucess) {
+                        return cadeira
+                    } else {
+                        return null
+                    }
+                })               
+        );
     }
-    return result_check;
-  }
-  
-  checkCriar(cadeira: Cadeira): string {
-    if (!this.cadeiraNaoCadastrada(cadeira.nome_disciplina)) {
-      return "Cadeira Já Cadastrada";
-    } if (!this.validNomeDisc(cadeira.nome_disciplina)) {
-      return "Nome da Disciplina Inválido";
-    } if (!this.validNomeProf(cadeira.nome_professor)) {
-      return "Nome de Professor Inválido";
-    } if (!this.validNumVagas(cadeira.numero_vagas)) {
-      return "Numero de Vagas Inválido";
-    } if (!this.validCargHoraria(cadeira.carga_horaria)) {
-      return "Horário inválido";
-    } if (!this.validHorarios(cadeira)) {
-      return "Estes horários já foram preenchidos, tente novamente";
+
+    atualizar(cadeira: Cadeira): Observable<Cadeira> {
+        return this.http.put<any>(this.taURL + "/cadeira", JSON.stringify(cadeira), {headers: this.headers})
+          .pipe( 
+            retry(2),
+            map( res => {
+                if (res.success) {
+                    return cadeira;
+                } else {
+                    return null;
+                }
+            })
+        ); 
     }
-    return "ok";
-  }
 
-  validHorarios(novaCadeira: Cadeira): boolean {
-    for (let cadeira of this.cadeiras){
-      if (cadeira.nome_professor == novaCadeira.nome_professor){
-        for (var key in cadeira.horarios){ 
-          var alocacaoAtual = cadeira.horarios[key];
-          var novaAlocacao = novaCadeira.horarios[key];
-          if (equal(alocacaoAtual, novaAlocacao) || interseccao(alocacaoAtual, novaAlocacao)) {
-            return false;
-          }
-        }
-      }
+    getCadeiras(departamento_ofertante: String = ""): Cadeira[] {
+        var result: Cadeira[] = [];
+        this.http.get<Cadeira[]>(this.taURL + "/cadeiras")
+                            .pipe(
+                                retry(2)
+                            ).subscribe(
+                                ar => {
+                                        if (ar) {
+                                            for (let c of ar) {
+                                                if (departamento_ofertante=="" || c.departamento_ofertante==departamento_ofertante) {
+                                                  result.push(c.clone());
+                                                }
+                                            }
+                                        }
+                                    },
+                                msg => { alert(msg.message); }
+                            );
+        return result;        
     }
-    return true;
-  }
-  
-  validCargHoraria(num: number): boolean {
-    if (!Number.isInteger(num) || num <= 0) {
-      return false;
+
+    getDepartamentos(): Observable<string[]> {
+        return this.http.get<string[]>(this.taURL + "/departamentos")
+            .pipe(
+                retry(2)
+            )
     }
-    return true;
-  }
 
-  validNomeDisc(nome_disciplina: string): boolean {
-    if (nome_disciplina.replace(/\s/g, '') == "") {
-      return false;
-    }
-    return true;
-  }
-
-  validNomeProf(nome_professor: string): boolean {
-    if (nome_professor.replace(/\s/g, '') == "") {
-      return false;
-    }
-    return true;
-  }
-
-  validNumVagas(num_str: string) {
-    if (Number.isNaN(num_str)) {
-      return false;
-    } 
-    const num = Number(num_str);
-    if (!Number.isInteger(num) || num <= 0) {
-      return false;
-    }
-    return true;
-  }
-
-  cadeiraNaoCadastrada(nome_disciplina: string): boolean {
-    return !this.cadeiras.find(a => a.nome_disciplina == nome_disciplina);
-  }
-
-  atualizarAlunos(cadeira: Cadeira): void {
-    cadeira = cadeira.clone();
-    for (let c of this.cadeiras)
-      if (c.nome_disciplina == cadeira.nome_disciplina) {
-        c.alunos = cadeira.alunos;
-      }
-  }
-  
-  getCadeiras(departamento_ofertante=""): Cadeira[] {
-    var result: Cadeira[] = [];
-    for (let c of this.cadeiras) {
-      if (departamento_ofertante=="" || c.departamento_ofertante==departamento_ofertante) {
-        result.push(c.clone());
-      }
-    }
-    return result;
-  }
-
-  getCadeira(nome: string): Cadeira | null {
-    var result = this.cadeiras.find(a => {
-      a.nome_disciplina == nome;
-    })
-    return result;
-  }
-
-  getDepartamentos(): string[] {
-    var result: string[] = [];
-    this.departamentos.forEach(d => {
-      result.push(d);
-    })
-    return result;
-  }
-
-  getTable(): Map<string,Cadeira[]> {
-    var result: Map<string,Cadeira[]> = new Map<string,Cadeira[]>();
-    this.departamentos.forEach(d => {
-      result.set(d, this.getCadeiras(d)) ;
-    })
-    return result;
-  }
-
-  addAluno(cadeira: Cadeira, aluno: Aluno): boolean {
-    cadeira = cadeira.clone();
-    var result = false;
-    for (let c of this.cadeiras) {
-      if (c.nome_disciplina == cadeira.nome_disciplina) {
-        result = c.addAluno(aluno);
+    getCadeira(nome: string): Cadeira | null {
+        var result = null;
+        this.http.get<Cadeira[]>(this.taURL + "/cadeiras")
+            .pipe(
+                retry(2)
+            ).subscribe(
+                ar => {
+                        if (ar) {
+                            for (let c of ar) {
+                                if (c.nome_disciplina == nome) {
+                                result.push(c.clone());
+                                break;
+                                }
+                            }
+                        }
+                    },
+                msg => { alert(msg.message); }
+            );
         return result;
-      }
     }
-    return result;
-  }
+
+    addAluno(cadeira: Cadeira, aluno: Aluno): Observable<Cadeira> {
+        var cadeiraAluno = new CadeiraAluno(cadeira, aluno)
+        return this.http.put<any>(this.taURL + "/cadeiraAddAluno", JSON.stringify(cadeiraAluno), {headers: this.headers})
+                .pipe(
+                    retry(2),
+                    map( res => {
+                        if (res.success) {
+                            return cadeira;
+                        } else {
+                            return null;
+                        }
+                    })
+                );
+    }
 }
-
-
-function equal(arg0: Set<number>, arg1: Set<number>) {
-  if (arg0.size !== arg1.size) return false;
-  if (arg0.size == 0 &&  arg1.size == 0) return false;
-  for (var a of arg0) if (!arg1.has(a)) return false;
-  return true;
-}
-
-function interseccao(setA: Set<number>, setB: Set<number>): boolean {
-  var _interseccao = new Set();
-  for (var elem of setB) {
-      if (setA.has(elem)) {
-          _interseccao.add(elem);
-      }
-  }
-  return _interseccao.size !== 0;
-}
-
